ajuste final

+++ user-service/internal/service/user-service.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"errors"
+	"fmt"
 	"time"
 	"user-service/internal/domain/model"
 	"user-service/internal/domain/ports"
@@ -36,28 +37,62 @@ func (s *UserService) Login(email, password string) (string, error) {
 	return token, err
 }
 
-func (s *UserService) RecoverPassword(email string) error {
-	return s.Repo.RecoverPassword(email)
+func (s *UserService) RecoverPassword(email string) (string, error) {
+    user, err := s.Repo.GetByEmail(email)
+    if err != nil {
+        return "", err
+    }
+
+    // De momento solo usamos el ID en la URL (más adelante podemos usar un token seguro)
+    recoveryURL := fmt.Sprintf("http://localhost:8080/user/password/%d", user.ID)
+
+    return recoveryURL, nil
 }
+
+func (s *UserService) UpdatePassword(id uint, password string) error {
+    user, err := s.Repo.GetId(int(id))
+    if err != nil {
+        return err
+    }
+
+    hashPassword, _ := hash.HashPassword(password)
+    user.Password = hashPassword
+
+    return s.Repo.Update(user)
+}
+
+
 
 func (s *UserService) GetId(userId int) (*model.User, error) {
 	return s.Repo.GetId(userId)
 }
 
-func (s *UserService) Update(email, password string) error {
-	user, err := s.Repo.GetByEmail(email)
+func (s *UserService) Update(user *model.User) error {
+	userdb, err := s.Repo.GetId(int(user.ID))
 	if err != nil {
 		return err
 	}
-	hashPassword, _ := hash.HashPassword(password)
-	user.Password = hashPassword
-	return s.Repo.Update(user)
+
+	comparedSucces := hash.ComparePassword(userdb.Password, user.Password)
+	if comparedSucces != nil {
+		return errors.New("invalid password")
+	}
+
+    return s.Repo.Update(user)
 }
 
-func (s *UserService) Delete(email string) error {
-	user, err := s.Repo.GetByEmail(email)
+
+func (s *UserService) Delete(userId int, email string, password string) error {
+	user, err := s.Repo.GetId(userId)
 	if err != nil {
 		return err
+	}
+	if user.Email != email || email != "admin@admin.com" {
+		return errors.New("invalid email")
+	}
+	comparedSucces := hash.ComparePassword(user.Password, password)
+	if comparedSucces != nil || password != "admin@admin.com" {
+		return errors.New("invalid password")
 	}
 	return s.Repo.Delete(user)
 }
